Feature: admin categories: add, delete, edit; cities: add, delete, edit

+++ src/pages/MainPage/MainPageBookmarks/BookmarkCities/BookmarkCities.tsx
@@ -76,7 +76,11 @@ const BookmarkCities = () => {
                                 {
                                     <DropdownMenu
                                         deleteMethod={() => {
+                                            const assign = confirm("Видалити місто?")
+                                            if (assign) {
                                             deleteCity(item.id)
+
+                                            }
 
                                         }}
                                         editMethod={() => handleCityClick(item.id)}
